Made prefixes 32-bit, improved vanilla IO compatibility.

+++ patches/tModLoader/Terraria/ModLoader/ModContent.cs
@@ -320,7 +320,7 @@ namespace Terraria.ModLoader
 		/// <summary>
 		/// Get the id (type) of a ModPrefix by class. Assumes one instance per class.
 		/// </summary>
-		public static byte PrefixType<T>() where T : ModPrefix => GetInstance<T>()?.Type ?? 0;
+		public static int PrefixType<T>() where T : ModPrefix => GetInstance<T>()?.Type ?? 0;
 
 		/// <summary>
 		/// Get the id (type) of a ModRarity by class. Assumes one instance per class.
